added comments to secret shares, added imports and uncommented tests in end_to_end_decentralized,added comments in training_information, potential fix for round_tracker
